Move all unit tests using mockito to mimic ESI endpoints to integration tests
