+++ tests/character/mod.rs
@@ -1,0 +1,2 @@
+mod get_character;
+mod get_character_affiliation;
