+++ tests/corporation/mod.rs
@@ -1,0 +1,1 @@
+mod get_corporation;
