+++ tests/tests.rs
@@ -1,0 +1,3 @@
+mod alliance;
+mod character;
+mod corporation;
